优化命名

+++ service-DigitalHuman/src/main/java/com/sugarweb/digitalHuman/application/KbService.java
@@ -1,5 +1,6 @@
 package com.sugarweb.digitalHuman.application;
 
+import cn.hutool.core.util.IdUtil;
 import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
@@ -8,8 +9,15 @@ import com.sugarweb.digitalHuman.application.dto.KbDetailDto;
 import com.sugarweb.digitalHuman.application.dto.KbPageQuery;
 import com.sugarweb.digitalHuman.application.dto.KbSaveDto;
 import com.sugarweb.digitalHuman.application.dto.KbUpdateDto;
+import com.sugarweb.digitalHuman.config.ChatAssistantProperties;
 import com.sugarweb.digitalHuman.domain.KbInfo;
+import com.sugarweb.digitalHuman.domain.ModelInfo;
+import com.sugarweb.digitalHuman.infra.MilvusEmbeddingStoreFactory;
+import com.sugarweb.framework.exception.ValidateException;
 import com.sugarweb.framework.orm.PageHelper;
+import dev.langchain4j.store.embedding.milvus.MilvusEmbeddingStore;
+import jakarta.annotation.Resource;
+import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 
@@ -19,7 +27,14 @@ import java.time.LocalDateTime;
  * @author xxd
  * @version 1.0
  */
+@Service
 public class KbService {
+
+    @Resource
+    private ChatAssistantProperties.MilvusVectorStoreProperties vectorStoreProperties;
+    @Resource
+    private MilvusEmbeddingStore embeddingStore;
+
     public IPage<KbDetailDto> page(KbPageQuery query) {
         return Db.page(PageHelper.getPage(query), new LambdaQueryWrapper<>(KbInfo.class)
                 .like(StrUtil.isNotEmpty(query.getKbName()), KbInfo::getKbName, query.getKbName())
@@ -34,11 +49,23 @@ public class KbService {
     public KbDetailDto save(KbSaveDto saveDto) {
         KbInfo kbInfo = new KbInfo();
         kbInfo.setKbName(saveDto.getKbName());
-        kbInfo.setEmbeddingModelId(saveDto.getEmbeddingModel());
         kbInfo.setStatus(saveDto.getStatus());
         kbInfo.setDescription(saveDto.getDescription());
         kbInfo.setCreateTime(LocalDateTime.now());
         kbInfo.setUpdateTime(LocalDateTime.now());
+
+        //绑定向量模型
+        ModelInfo modelInfo = Db.getById(saveDto.getEmbeddingModelId(), ModelInfo.class);
+        if (modelInfo == null) {
+            throw new ValidateException("模型未配置");
+        }
+        kbInfo.setEmbeddingModelId(saveDto.getEmbeddingModelId());
+        kbInfo.setEmbeddingModelName(modelInfo.getModelName());
+        kbInfo.setDimension(modelInfo.getDimension());
+
+        //绑定向量库
+        bindVectorDatabase(kbInfo);
+
         Db.save(kbInfo);
         return buildKbDetailDto(kbInfo);
     }
@@ -47,10 +74,17 @@ public class KbService {
         KbInfo kbInfo = Db.getById(updateDto.getKbId(), KbInfo.class);
         if (kbInfo != null) {
             kbInfo.setKbName(updateDto.getKbName());
-            kbInfo.setEmbeddingModelId(updateDto.getEmbeddingModel());
             kbInfo.setStatus(updateDto.getStatus());
             kbInfo.setDescription(updateDto.getDescription());
             kbInfo.setUpdateTime(LocalDateTime.now());
+
+            ModelInfo modelInfo = Db.getById(updateDto.getEmbeddingModelId(), ModelInfo.class);
+            if (modelInfo == null) {
+                throw new ValidateException("模型未配置");
+            }
+            kbInfo.setEmbeddingModelId(updateDto.getEmbeddingModelId());
+            kbInfo.setEmbeddingModelName(modelInfo.getModelName());
+            kbInfo.setDimension(modelInfo.getDimension());
             Db.updateById(kbInfo);
         }
         return buildKbDetailDto(kbInfo);
@@ -73,8 +107,20 @@ public class KbService {
         return kbDetailDto;
     }
 
-
     public void remove(String kbId) {
+        KbInfo kbInfo = Db.getById(kbId, KbInfo.class);
+        if (kbInfo == null) {
+            throw new ValidateException("知识库不存在");
+        }
+        embeddingStore.dropCollection(kbInfo.getCollectionName());
         Db.removeById(kbId, KbInfo.class);
     }
+
+    private void bindVectorDatabase(KbInfo kbInfo) {
+        String collectionName = IdUtil.randomUUID();
+        kbInfo.setCollectionName(collectionName);
+
+        MilvusEmbeddingStoreFactory.create(kbInfo.getCollectionName(), kbInfo.getDimension());
+    }
+
 }
